requirements.txt
tensorflow>=2.20
numpy
Pillow
gradio
matplotlib